printer: implement "print_number" method
It takes a pointer to a JSON type as input, then access
the element of the JSON struct which stores parsed number.
The parsed number is then populated in a string buffer
which is returned by this method.

The string buffer is stored in a dynamic memory which needs to
be free'd after printing.

NB: This method tries to deal with integer, float etc. but the
precision might be off in some cases.